+++ ESFramework/Core/Additions/NSDictionary+ESAdditions.h
@@ -24,14 +24,3 @@
 - (NSDictionary<KeyType, ObjectType> *)entriesWithOptions:(NSEnumerationOptions)opts passingTest:(BOOL (^)(KeyType key, ObjectType obj, BOOL *stop))predicate;
 
 @end
-
-@interface NSMutableDictionary (ESAdditions)
-
-/**
- * Set the value for the derived property identified by a given key path.
- *
- * @param keyPath A key path of the form relationship.property (with one or more relationships); for example “department.name” or “department.manager.lastName”.
- */
-- (BOOL)es_setValue:(id)value forKeyPath:(NSString *)keyPath;
-
-@end