+++ ESFramework/Core/Additions/NSDictionary+ESAdditions.m
@@ -8,9 +8,6 @@
 
 #import "NSDictionary+ESAdditions.h"
 #import "NSURLComponents+ESAdditions.h"
-#import "ESDefines.h"
-#import "ESValue.h"
-#import "NSString+ESAdditions.h"
 
 @implementation NSDictionary (ESAdditions)
 
@@ -36,52 +33,6 @@
     NSArray *keys = [self keysOfEntriesWithOptions:opts passingTest:predicate].allObjects;
     NSArray *objects = [self objectsForKeys:keys notFoundMarker:NSNull.null];
     return [NSDictionary dictionaryWithObjects:objects forKeys:keys];
-}
-
-@end
-
-@implementation NSMutableDictionary (ESAdditions)
-
-- (BOOL)es_setValue:(id)value forKeyPath:(NSString *)keyPath
-{
-    if (![keyPath isKindOfClass:[NSString class]]) {
-        return NO;
-    }
-    if (![keyPath contains:@"."]) {
-        self[keyPath] = value;
-        return YES;
-    }
-    NSArray *keys = [keyPath componentsSeparatedByString:@"."];
-    if (keys.count < 2) {
-        return NO;
-    }
-
-    NSUInteger maxIndex = keys.count - 1;
-    __block NSMutableString *currentKeyPath = [NSMutableString string];
-    __block NSMutableDictionary *mutableCopy = self.mutableCopy;
-    [keys enumerateObjectsUsingBlock:^(id subKey, NSUInteger idx, BOOL *stop) {
-        [currentKeyPath appendFormat:@"%@%@", (currentKeyPath.length > 0 ? @"." : @""), subKey];
-        if (idx == maxIndex) {
-            [mutableCopy setValue:value forKeyPath:currentKeyPath];
-            return;
-        }
-        id currentValue = [mutableCopy valueForKeyPath:currentKeyPath];
-        if (!currentValue) {
-            [mutableCopy setValue:[NSMutableDictionary dictionary] forKeyPath:currentKeyPath];
-        } else if ([currentValue isKindOfClass:[NSDictionary class]] &&
-                   ![currentValue isKindOfClass:[NSMutableDictionary class]])
-        {
-            [mutableCopy setValue:[currentValue mutableCopy] forKeyPath:currentKeyPath];
-        } else if (![currentValue isKindOfClass:[NSDictionary class]]) {
-            mutableCopy = nil;
-            *stop = YES;
-        }
-    }];
-    if (mutableCopy) {
-        [self setDictionary:[mutableCopy copy]];
-        return YES;
-    }
-    return NO;
 }
 
 @end